fix: replace portal Modal by Modal of headlessui
